+++ pyannote/audio/utils/protocol.py
@@ -20,7 +20,9 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
+from functools import partial
 
+import torchaudio
 from pyannote.database import FileFinder, Protocol, get_annotated
 from pyannote.database.protocol import SpeakerVerificationProtocol
 
@@ -89,7 +91,14 @@ def check_protocol(protocol: Protocol) -> Protocol:
 
     if "waveform" not in file and "torchaudio.info" not in file:
 
-        protocol.preprocessors["torchaudio.info"] = get_torchaudio_info
+        # use soundfile when available (it usually is faster than ffmpeg for getting info)
+        backends = (
+            torchaudio.list_audio_backends()
+        )  # e.g ['ffmpeg', 'soundfile', 'sox']
+        backend = "soundfile" if "soundfile" in backends else backends[0]
+        protocol.preprocessors["torchaudio.info"] = partial(
+            get_torchaudio_info, backend=backend
+        )
         msg = (
             f"Protocol {protocol.name} does not precompute the output of torchaudio.info(): "
             f"adding a 'torchaudio.info' preprocessor for you to speed up dataloaders. "